+++ setup.php
@@ -35,7 +35,7 @@ use GlpiPlugin\Datainjection\Glpi\Asset\Capacity\IsInjectableCapacity;
 use function Safe\define;
 use function Safe\mkdir;
 
-define('PLUGIN_DATAINJECTION_VERSION', '2.15.8');
+define('PLUGIN_DATAINJECTION_VERSION', '2.15.10');
 
 // Minimal GLPI version, inclusive
 define("PLUGIN_DATAINJECTION_MIN_GLPI", "11.0.5");